:boar: perf: Improve storage access using unsafe pointers
Allow old usage by setting a build-tag=nounsafe
